fix #2283 - migrate trustedshops to new database schema
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@14813 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ admin/includes/modules/system/trustedshops.php
@@ -42,12 +42,10 @@ class trustedshops {
       );
     
       foreach ($table_array as $table) {
-        if (in_array($table['column'], $db_table_rows)) {
-          if ($table['column'] == 'add') {
-            xtc_db_query("ALTER TABLE ".TABLE_TRUSTEDSHOPS." ADD ".$table['column']." ".$table['default']."");
-          } else {
-            xtc_db_query("ALTER TABLE ".TABLE_TRUSTEDSHOPS." DROP COLUMN ".$table['column']);
-          }
+        if (!in_array($table['column'], $db_table_rows) && $table['column'] == 'add') {
+          xtc_db_query("ALTER TABLE ".TABLE_TRUSTEDSHOPS." ADD ".$table['column']." ".$table['default']."");
+        } elseif (in_array($table['column'], $db_table_rows) && $table['column'] == 'delete') { {
+          xtc_db_query("ALTER TABLE ".TABLE_TRUSTEDSHOPS." DROP COLUMN ".$table['column']);
         }
       }
     }
